Avoid to associate an id to COSNull when writing.
This was leading to a case where COSNull was given an id and associated
to the an indirect reference. The indirect reference value was later set
a value (a COSInteger representing a objects stream length). Subsequent
calls to getOrCreateIndirectReferenceFor(COSNull.NULL) where returning
the reference to the COSInteger which was written where "null" was
expected to be written instead